Keep location.query in sync with state
The reason for this is because viewing a specific plugin is a separate
component which results in the Main component being unmounted. So after
closing the PluginDetail modal, browserHistory.goBack() is called. But
since the component was unmounted its state is lost. When Main is
mounted it reads location.query for any given values.