Save state between components
Because we have a requirement of needing external URLs for seeing a
plugin and its details (<host>/<plugin name>) we need to able to save
the current search criteria so when returning to the search we have the
same screen as before.

I tried various combinations of location.query syncing and
location.state and couldn't find anything that worked. So this hack
stores the Main component state at componentWillUnmount and reads it
back during componentWillMount, where it also then reads any query
params to keep parody with the existing implementation.